backend/requirements.txt
anyio==4.4.0
backoff==2.2.1
black==24.4.2
boto3==1.34.128
botocore==1.34.128
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
gql==3.5.0
graphql-core==3.2.3
idna==3.7
jmespath==1.0.1
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
tomli==2.0.1
typing_extensions==4.12.0
urllib3==1.26.19
yarl==1.9.4
mlflow==2.20.3 
sagemaker-mlflow==0.1.0
langgraph==0.3.5
